pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardrail-eval"
version = "0.1.0"
description = "Unified evaluation harness for multimodal guardrail models"
requires-python = ">=3.11"
dependencies = [
    "vllm>=0.11.0",
    "transformers>=4.57.1",
    "datasets>=3.0.0",
    "huggingface_hub>=0.26.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.6.0"]

[project.scripts]
guardrail-eval = "guardrail_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
